Spring Initializr/bilibili-analysis/python_scripts/requirements.txt
bilibili-api-python
pymysql
aiohttp